added separate files, tests, precommit checks
